fix(ui): three rendering bugs surfaced by manual review
Bug 1 — legacy investment Valuations panel created a huge gap
between the chart and the transactions list (full-content
~900pt height for 60+ valuations). Wrap the panel's macOS body in
`ScrollView { VStack(Divider-separated) }` for internal scroll, and
bound the height at the call site in `legacyChartAndValuations`
macOS branch to ~340pt (matches the chart's natural height:
time-period picker + chart + padding).

Bug 2 + 3 — position-tracked investment AND multi-currency
standard accounts rendered only the Grid's Instrument column,
hiding the other five (Qty, Unit Price, Cost, Value, Gain). Root
cause: `PositionsTableRow` was a custom `View` struct wrapping
`GridRow { ... 6 cells ... }`, but SwiftUI's `Grid` cannot
introspect through custom `View` wrappers to find their inner
`GridRow` for column alignment — it treats the wrapper as a
single-column child. Inline the `GridRow` directly inside
`macOSGridLayout`'s `ForEach`, moving cell helpers and the
`rowBackground` semantic-token helper to a `private` extension on
`PositionsTable`. Per-row `@State isHovered` becomes a single
`@State hoveredRowIndex: Int?` driven by the `.onHover` modifiers
— SwiftUI diffs per-row from the index comparison. Delete
`Shared/Views/Positions/PositionsTableRow.swift`.

Bug 2 — chart collapsed to its minimum intrinsic height inside
the unbounded vertical `List` row. Spec Risk #3 explicitly
authorises a localised `.frame(minHeight: 220)` on `PositionsChart`
once the collapse is observed in the running app — applied at
PositionsChart's root so every embedding benefits.
