+++ Shared/Views/Positions/PositionsTable+GridLayout.swift
@@ -7,6 +7,16 @@ import SwiftUI
   /// `file_length` / `type_body_length` thresholds; the `@State`
   /// properties driving this code path stay on the main type (Swift
   /// disallows stored properties in extensions).
+  ///
+  /// `GridRow`s for body rows are inlined directly inside the outer
+  /// `Grid { ForEach { … } }` rather than wrapped in a custom `View`
+  /// struct. SwiftUI's `Grid` only walks its **direct** children when
+  /// computing column alignment; nesting `GridRow` inside a custom
+  /// `View` (e.g. `struct PositionsTableRow: View { var body: some
+  /// View { GridRow { … } } }`) makes the parent `Grid` see each
+  /// wrapper as a single-column child, which collapsed the visible
+  /// table to its first column only (Bugs #2/#3 surfaced by manual
+  /// review). Inlining the `GridRow` restores column alignment.
   extension PositionsTable {
     // MARK: - macOS Grid
 
@@ -20,16 +30,18 @@ import SwiftUI
         headerRow
         Divider().gridCellColumns(6)
         ForEach(Array(sortedRows.enumerated()), id: \.element.id) { item in
-          // Closure parameter tuple destructuring (`{ idx, row in }`)
-          // was removed in Swift 4 (SE-0110); destructure inside the
-          // body.
-          let (idx, row) = (item.offset, item.element)
-          PositionsTableRow(
-            row: row,
-            isSelected: selection?.id == row.id,
-            isFocused: focusedRowIndex == idx && isPanelFocused,
-            isAlternateRow: !idx.isMultiple(of: 2),
-            toggleSelection: { toggleSelection(row.instrument) })
+          // `GridRow` is inlined here (not extracted to a `struct
+          // PositionsTableRow: View`) because SwiftUI's `Grid` does
+          // not introspect through custom-View `body` boundaries to
+          // find inner `GridRow`s for column alignment — wrapping
+          // each row in a custom View collapsed the table to its
+          // first column (manual-review Bugs #2/#3). Chrome
+          // (padding/background/focus/tap/hover) is applied via
+          // `.modifier(PositionsGridRowChrome…)` so the result is
+          // `ModifiedContent<GridRow, _>` rather than a separate
+          // `body` boundary; modifiers attached directly to a
+          // `GridRow` remain visible to the Grid layout engine.
+          gridBodyRow(item: item)
         }
       }
       .padding(.horizontal, 12)
@@ -103,6 +115,49 @@ import SwiftUI
       selection = (selection?.id == instrument.id) ? nil : instrument
     }
 
+    /// Builds one body `GridRow` with cells inline and chrome applied
+    /// via `.modifier(PositionsGridRowChrome…)`. The function returns
+    /// `some View`, not a `struct: View`, so the call site at
+    /// `ForEach { gridBodyRow(item:) }` resolves to
+    /// `ModifiedContent<GridRow, _>` — preserving Grid's ability to
+    /// see the inner row for column alignment (custom View wrappers
+    /// would break that, see Bugs #2/#3).
+    @ViewBuilder
+    private func gridBodyRow(item: EnumeratedSequence<[ValuedPosition]>.Element)
+      -> some View
+    {
+      // Closure parameter tuple destructuring (`{ idx, row in }`) was
+      // removed in Swift 4 (SE-0110); destructure inside the body.
+      let (idx, row) = (item.offset, item.element)
+      GridRow(alignment: .firstTextBaseline) {
+        gridInstrumentCell(for: row)
+        Text(row.quantityFormatted)
+          .monospacedDigit()
+          .gridColumnAlignment(.trailing)
+        gridAmountCell(row.unitPrice).gridColumnAlignment(.trailing)
+        gridAmountCell(row.costBasis).gridColumnAlignment(.trailing)
+        gridAmountCell(row.value).gridColumnAlignment(.trailing)
+        gridGainCell(for: row)
+          .gridColumnAlignment(.trailing)
+          .frame(minWidth: 140, alignment: .trailing)
+      }
+      .modifier(
+        PositionsGridRowChrome(
+          isSelected: selection?.id == row.id,
+          isHovered: hoveredRowIndex == idx,
+          isWindowKey: controlActiveState == .key,
+          isAlternateRow: !idx.isMultiple(of: 2),
+          isFocused: focusedRowIndex == idx && isPanelFocused,
+          onTap: { toggleSelection(row.instrument) },
+          onHoverChange: { hovering in
+            if hovering {
+              hoveredRowIndex = idx
+            } else if hoveredRowIndex == idx {
+              hoveredRowIndex = nil
+            }
+          }))
+    }
+
     // MARK: - Header row (visible Grid headers)
 
     @ViewBuilder private var headerRow: some View {
@@ -151,6 +206,59 @@ import SwiftUI
       // activatable.
       .buttonStyle(.borderless)
       .gridColumnAlignment(alignment)
+    }
+
+    // MARK: - Body-row cells (inlined from former `PositionsTableRow`)
+
+    /// Instrument cell with kind badge + name + optional exchange.
+    /// Named `gridInstrumentCell` to disambiguate from the iOS Table
+    /// path's `instrumentCell(for:)` in the private extension on the
+    /// main file.
+    @ViewBuilder
+    private func gridInstrumentCell(for row: ValuedPosition) -> some View {
+      HStack(spacing: 6) {
+        KindBadge(kind: row.instrument.kind)
+        VStack(alignment: .leading) {
+          Text(row.instrument.name)
+          if let exchange = row.instrument.exchange {
+            Text(exchange).font(.caption).foregroundStyle(.secondary)
+          }
+        }
+      }
+    }
+
+    /// Monetary-amount cell with `monospacedDigit()` and an em-dash
+    /// fallback when the amount is unavailable (e.g. conversion
+    /// failure).
+    @ViewBuilder
+    private func gridAmountCell(_ amount: InstrumentAmount?) -> some View {
+      if let amount {
+        Text(amount.formatted).monospacedDigit()
+      } else {
+        Text("—").foregroundStyle(.tertiary)
+      }
+    }
+
+    /// Gain cell with signed amount + percent suffix. Both pieces
+    /// share `gain.gainColor` so a red-leaning percent matches its
+    /// red-leaning dollar amount.
+    @ViewBuilder
+    private func gridGainCell(for row: ValuedPosition) -> some View {
+      if let gain = row.gainLoss {
+        HStack(spacing: 4) {
+          Text(gain.signedFormatted)
+            .monospacedDigit()
+            .foregroundStyle(gain.gainColor)
+          if let pct = row.gainLossPercent {
+            Text(GainLossPercentDisplay.formatted(pct))
+              .font(.caption)
+              .monospacedDigit()
+              .foregroundStyle(gain.gainColor)
+          }
+        }
+      } else {
+        Text("—").foregroundStyle(.tertiary)
+      }
     }
 
     // MARK: - Accessibility representation bindings
@@ -234,6 +342,56 @@ import SwiftUI
     private func accessibilityGainText(for row: ValuedPosition) -> String {
       guard let gain = row.gainLoss else { return "no gain or loss" }
       return gainAccessibilityLabel(gain: gain, percent: row.gainLossPercent)
+    }
+  }
+
+  /// Row chrome (background, focus ring, hover/tap handling, padding)
+  /// applied to a body `GridRow` via `.modifier(_:)`. Implemented as a
+  /// `ViewModifier` rather than a custom `View` wrapper because a
+  /// `ViewModifier` returns `some View` from its `body(content:)`
+  /// without introducing a new view-tree boundary: the result is
+  /// `ModifiedContent<GridRow, PositionsGridRowChrome>`, which the
+  /// outer `Grid` still sees as a `GridRow` for column-alignment
+  /// purposes. A `struct: View` wrapper would create a separate `body`
+  /// boundary that `Grid` cannot introspect (manual-review Bugs
+  /// #2/#3).
+  private struct PositionsGridRowChrome: ViewModifier {
+    let isSelected: Bool
+    let isHovered: Bool
+    let isWindowKey: Bool
+    let isAlternateRow: Bool
+    let isFocused: Bool
+    let onTap: () -> Void
+    let onHoverChange: (Bool) -> Void
+
+    func body(content: Content) -> some View {
+      content
+        .padding(.vertical, 6)
+        .padding(.horizontal, 8)
+        .background(background)
+        .overlay(focusOverlay)
+        .contentShape(Rectangle())
+        .onTapGesture(perform: onTap)
+        .onHover(perform: onHoverChange)
+    }
+
+    private var background: Color {
+      if isSelected {
+        return isWindowKey
+          ? Color(nsColor: .selectedContentBackgroundColor)
+          : Color(nsColor: .unemphasizedSelectedContentBackgroundColor)
+      }
+      if isHovered {
+        return Color(nsColor: .controlAccentColor).opacity(0.10)
+      }
+      return Color(nsColor: NSColor.alternatingContentBackgroundColors[isAlternateRow ? 1 : 0])
+    }
+
+    @ViewBuilder private var focusOverlay: some View {
+      if isFocused {
+        RoundedRectangle(cornerRadius: 4)
+          .stroke(Color.accentColor, lineWidth: 2)
+      }
     }
   }
